comply with ISO C++
replaced unsigned long long by uint64_t

+++ pkg/src/nabc_globals.h
@@ -12,11 +12,10 @@
 
 typedef double flt_ty; /**< type for scientific accuracy in rate constants */
 #ifdef USE_UNSIGNEDLONG
-	typedef unsigned long int ulint;/**< type for population counts */
+typedef unsigned long int ulint;/**< type for population counts */
 #else
-typedef unsigned long long ulint;/**< type for population counts */
+typedef uint64_t ulint;/**< type for population counts */
 
-//typedef uint64_t ulint;/**< type for population counts */
 #endif
 typedef unsigned int uint;
